Moved the nldi data reducers and actions into store/nldi-data.js and refactoring necessary for this change.
Also created a mock-service-data module so mock service responses can be shared across test modules
